apps/api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.20.0
psycopg2-binary>=2.9.9
alembic>=1.13.1
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
redis>=5.0.1
rq>=1.15.1
httpx>=0.26.0
google-api-python-client>=2.114.0
google-auth-oauthlib>=1.2.0
openai>=1.10.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.7
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.2.1
ffmpeg-python>=0.2.0
pillow>=10.2.0
scikit-learn>=1.4.0
numpy>=1.26.3
yt-dlp>=2024.08.06
greenlet>=3.0.0
